+++ CHANGELOG.rst
@@ -6,6 +6,15 @@ veeamhub.veeam Release Notes
 
 This changelog describes changes after version 1.5.0.
 
+v2.3.8
+======
+
+Minor Changes
+-------------
+
+- Added new module - `defender_exclusions` - enabling management of Microsoft Defender exclusions including playbook examples of setting Veeam exclusions.
+- Updated documentation of all modules in this collection to use FQCN in all examples.
+
 v2.3.7
 ======